+++ src/asymmetric_crypto/curve_25519/mod.rs
@@ -1,5 +1,6 @@
 mod curve_secret;
 mod ed25519;
+mod private_key;
 mod ristretto_25519;
 mod x25519;
 
@@ -10,5 +11,12 @@ pub use ed25519::{
     Cached25519Signer, Ed25519Keypair, Ed25519PrivateKey, Ed25519PublicKey, Ed25519Signature,
     ED25519_PUBLIC_KEY_LENGTH,
 };
-pub use ristretto_25519::{R25519PrivateKey, R25519PublicKey, R25519_PUBLIC_KEY_LENGTH};
-pub use x25519::{X25519Keypair, X25519PrivateKey, X25519PublicKey, X25519_PUBLIC_KEY_LENGTH};
+pub use private_key::CURVE_25519_PRIVATE_KEY_LENGTH;
+pub use ristretto_25519::{
+    R25519CurvePoint, R25519PrivateKey, R25519PublicKey, R25519_PRIVATE_KEY_LENGTH,
+    R25519_PUBLIC_KEY_LENGTH,
+};
+pub use x25519::{
+    X25519CurvePoint, X25519Keypair, X25519PrivateKey, X25519PublicKey, X25519_PRIVATE_KEY_LENGTH,
+    X25519_PUBLIC_KEY_LENGTH,
+};