+++ src/ecies/ecies_nist_old.rs
@@ -1,58 +1,61 @@
-use aes_gcm::Aes128Gcm as Aes128GcmLib;
+use aead::{consts::U10, generic_array::GenericArray};
+use chacha20::hchacha;
+use chacha20poly1305::XChaCha20Poly1305 as XChaCha20Poly1305Lib;
+use elliptic_curve::{Curve, CurveArithmetic};
 
 use super::ecies_traits::EciesStream;
 use crate::{
-    asymmetric_crypto::{R25519PrivateKey, R25519PublicKey},
-    kdf128,
+    blake2b,
     reexport::rand_core::CryptoRngCore,
-    symmetric_crypto::{Aes128Gcm, Dem, DemStream, Instantiable, Nonce, SymmetricKey},
-    CryptoCoreError, Ecies, FixedSizeCBytes, RandomFixedSizeCBytes,
+    symmetric_crypto::{Dem, DemStream, Instantiable, Nonce, SymmetricKey, XChaCha20Poly1305},
+    CryptoCoreError, Ecies, FixedSizeCBytes, NistPrivateKey, NistPublicKey, RandomFixedSizeCBytes,
 };
 
 /// A thread safe Elliptic Curve Integrated Encryption Scheme (ECIES) using
-///  - the Ristretto group of Curve 25519
-///  - AES 128 GCM
-///  - SHAKE256 (XOF)
-pub struct EciesR25519Aes128 {}
+///  - A NIST approved curve
+///  -`XChaCha20`
+///  - Blake2b
+pub struct EciesNistChaCha20 {}
 
-fn get_nonce<const NONCE_LENGTH: usize>(
-    ephemeral_pk: &R25519PublicKey,
-    recipient_pk: &R25519PublicKey,
-) -> Nonce<NONCE_LENGTH> {
+fn get_nonce<const NONCE_LENGTH: usize, C: Curve + CurveArithmetic>(
+    ephemeral_pk: &NistPublicKey<C>,
+    recipient_pk: &NistPublicKey<C>,
+) -> Result<Nonce<NONCE_LENGTH>, CryptoCoreError> {
     let mut nonce = Nonce([0; NONCE_LENGTH]);
-    kdf128!(
-        &mut nonce.0,
-        &ephemeral_pk.to_bytes(),
-        &recipient_pk.to_bytes()
+    blake2b!(nonce.0, ephemeral_pk.as_bytes(), recipient_pk.as_bytes())?;
+    Ok(nonce)
+}
+
+fn get_ephemeral_key<const KEY_LENGTH: usize, C: Curve + CurveArithmetic>(
+    shared_point: &NistPublicKey<C>,
+) -> Result<SymmetricKey<KEY_LENGTH>, CryptoCoreError> {
+    let key = hchacha::<U10>(
+        GenericArray::from_slice(shared_point.as_bytes()),
+        &GenericArray::default(),
     );
-    nonce
+
+    Ok(SymmetricKey(key.as_slice().try_into().map_err(|_| {
+        CryptoCoreError::InvalidBytesLength("get ephemeral key".to_string(), KEY_LENGTH, None)
+    })?))
 }
 
-fn get_ephemeral_key<const KEY_LENGTH: usize>(
-    shared_point: &R25519PublicKey,
-) -> SymmetricKey<KEY_LENGTH> {
-    let mut key = SymmetricKey([0; KEY_LENGTH]);
-    kdf128!(&mut key.0, &shared_point.to_bytes());
-    key
-}
-
-impl EciesR25519Aes128 {
-    fn generate_keys_and_nonce<R: CryptoRngCore>(
+impl EciesNistChaCha20 {
+    fn generate_keys_and_nonce<R: CryptoRngCore, C: Curve + CurveArithmetic>(
         rng: &mut R,
-        recipient_pk: &R25519PublicKey,
+        recipient_pk: &NistPublicKey<C>,
     ) -> Result<
         (
-            R25519PublicKey,
-            SymmetricKey<{ Aes128Gcm::KEY_LENGTH }>,
-            Nonce<{ Aes128Gcm::NONCE_LENGTH }>,
+            NistPublicKey<C>,
+            SymmetricKey<{ XChaCha20Poly1305::KEY_LENGTH }>,
+            Nonce<{ XChaCha20Poly1305::NONCE_LENGTH }>,
         ),
         CryptoCoreError,
     > {
-        let ephemeral_sk = R25519PrivateKey::new(rng);
-        let ephemeral_pk = R25519PublicKey::from(&ephemeral_sk);
+        let ephemeral_sk = NistPrivateKey::new(rng);
+        let ephemeral_pk = NistPublicKey::<C>::from(&ephemeral_sk);
 
         // Calculate the shared secret point (Px, Py) = P = r.Y
-        let shared_point = recipient_pk * &ephemeral_sk;
+        let shared_point = recipient_pk.dh(&ephemeral_sk);
 
         // Generate the 128-bit symmetric encryption key k, derived using SHAKE256
         // eXtendable-Output-Function (XOF) such as: k = kdf(S || S1) where:
@@ -60,26 +63,26 @@ impl EciesR25519Aes128 {
         // - S1: if the user provided shared_encapsulation_data S1, then we append it to
         //   the shared_bytes S
         // This implementation does NOT use the shared_encapsulation_data S1
-        let key = get_ephemeral_key::<{ Aes128Gcm::KEY_LENGTH }>(&shared_point);
+        let key = get_ephemeral_key::<{ XChaCha20Poly1305::KEY_LENGTH }>(&shared_point)?;
 
         // Calculate the nonce based on the 2 public keys
-        let nonce = get_nonce::<{ Aes128Gcm::NONCE_LENGTH }>(&ephemeral_pk, recipient_pk);
+        let nonce = get_nonce::<{ XChaCha20Poly1305::NONCE_LENGTH }>(&ephemeral_pk, recipient_pk)?;
 
         Ok((ephemeral_pk, key, nonce))
     }
 
-    fn recover_key_and_nonce(
-        recipient_sk: &R25519PrivateKey,
-        ephemeral_public_key: &R25519PublicKey,
+    fn recover_key_and_nonce<C: Curve + CurveArithmetic, const PRIVATE_KEY_LENGTH: usize>(
+        recipient_sk: &NistPrivateKey<C, PRIVATE_KEY_LENGTH>,
+        ephemeral_public_key: &NistPublicKey<C>,
     ) -> Result<
         (
-            SymmetricKey<{ Aes128Gcm::KEY_LENGTH }>,
-            Nonce<{ Aes128Gcm::NONCE_LENGTH }>,
+            SymmetricKey<{ XChaCha20Poly1305::KEY_LENGTH }>,
+            Nonce<{ XChaCha20Poly1305::NONCE_LENGTH }>,
         ),
         CryptoCoreError,
     > {
         // Calculate the shared secret point (Px, Py) = P = R.y = r.G.y = r.Y
-        let shared_point = ephemeral_public_key * recipient_sk;
+        let shared_point = ephemeral_public_key.dh(recipient_sk);
 
         // Generate the 128-bit symmetric encryption key k, derived using SHAKE256 XOF
         // such as: k = kdf(S || S1) where:
@@ -87,37 +90,39 @@ impl EciesR25519Aes128 {
         // - S1: if the user provided shared_encapsulation_data S1, then we append it to
         //   the shared_bytes S
         // This implementation does NOT use the shared_encapsulation_data S1
-        let key = get_ephemeral_key::<{ Aes128Gcm::KEY_LENGTH }>(&shared_point);
+        let key = get_ephemeral_key::<{ XChaCha20Poly1305::KEY_LENGTH }, C>(&shared_point)?;
 
         // Recompute the nonce
-        let nonce = get_nonce::<{ Aes128Gcm::NONCE_LENGTH }>(
+        let nonce = get_nonce::<{ XChaCha20Poly1305::NONCE_LENGTH }, C>(
             ephemeral_public_key,
-            &R25519PublicKey::from(recipient_sk),
-        );
+            &NistPublicKey::<C>::from(recipient_sk),
+        )?;
 
         Ok((key, nonce))
     }
 }
 
-impl Ecies<R25519PrivateKey, R25519PublicKey> for EciesR25519Aes128 {
-    const ENCRYPTION_OVERHEAD: usize = R25519PublicKey::LENGTH + Aes128Gcm::MAC_LENGTH;
+impl<C: Curve + CurveArithmetic, const PRIVATE_KEY_LENGTH: usize>
+    Ecies<NistPrivateKey<C, PRIVATE_KEY_LENGTH>, NistPublicKey<C>> for EciesNistChaCha20
+{
+    const ENCRYPTION_OVERHEAD: usize = NistPublicKey::<C>::LENGTH + XChaCha20Poly1305::MAC_LENGTH;
 
     fn encrypt<R: CryptoRngCore>(
         rng: &mut R,
-        recipient_pk: &R25519PublicKey,
+        recipient_pk: &NistPublicKey<C>,
         plaintext: &[u8],
         authentication_data: Option<&[u8]>,
     ) -> Result<Vec<u8>, CryptoCoreError> {
-        let (ephemeral_pk, key, nonce) = Self::generate_keys_and_nonce(rng, recipient_pk)
-            .map_err(|_| CryptoCoreError::EncryptionError)?;
+        let (ephemeral_pk, key, nonce) = Self::generate_keys_and_nonce(rng, recipient_pk)?;
 
         // Encrypt and authenticate the message, returning the ciphertext and MAC
         let ciphertext_plus_tag =
-            Aes128Gcm::new(&key).encrypt(&nonce, plaintext, authentication_data)?;
+            XChaCha20Poly1305::new(&key).encrypt(&nonce, plaintext, authentication_data)?;
 
-        // Assemble the final encrypted message: R || nonce || c || d
-        let mut res =
-            Vec::with_capacity(R25519PublicKey::LENGTH + plaintext.len() + Aes128Gcm::MAC_LENGTH);
+        // Assemble the final encrypted message: R || c || d
+        let mut res = Vec::with_capacity(
+            NistPublicKey::<C>::LENGTH + plaintext.len() + XChaCha20Poly1305::MAC_LENGTH,
+        );
         res.extend(ephemeral_pk.to_bytes());
         res.extend(&ciphertext_plus_tag);
 
@@ -125,72 +130,86 @@ impl Ecies<R25519PrivateKey, R25519PublicKey> for EciesR25519Aes128 {
     }
 
     fn decrypt(
-        recipient_sk: &R25519PrivateKey,
+        recipient_sk: &NistPrivateKey<C, PRIVATE_KEY_LENGTH>,
         ciphertext: &[u8],
         authentication_data: Option<&[u8]>,
     ) -> Result<Vec<u8>, CryptoCoreError> {
-        // Extract the sender's ephemeral public key R from the ciphertext
-        let ephemeral_pk = R25519PublicKey::try_from_slice(&ciphertext[..R25519PublicKey::LENGTH])?;
+        // Extract the ephemeral public key from the beginning of the ciphertext
+        let ephemeral_pk =
+            NistPublicKey::<C>::try_from_slice(&ciphertext[..NistPublicKey::<C>::LENGTH])?;
 
-        // Recover the key and nonce
+        // recover the symmetric key and nonce
         let (key, nonce) = Self::recover_key_and_nonce(recipient_sk, &ephemeral_pk)?;
 
         // Decrypt and verify the message using AES-128-GCM
-        let decrypted_message = Aes128Gcm::new(&key).decrypt(
+        XChaCha20Poly1305::new(&key).decrypt(
             &nonce,
-            &ciphertext[R25519PublicKey::LENGTH..],
+            &ciphertext[NistPublicKey::<C>::LENGTH..],
             authentication_data,
-        )?;
-
-        Ok(decrypted_message)
+        )
     }
 }
 
-impl EciesStream<R25519PrivateKey, R25519PublicKey, Aes128GcmLib> for EciesR25519Aes128 {
+impl<C: Curve + CurveArithmetic, const PRIVATE_KEY_LENGTH: usize>
+    EciesStream<NistPrivateKey<C, PRIVATE_KEY_LENGTH>, NistPublicKey<C>, XChaCha20Poly1305Lib>
+    for EciesNistChaCha20
+{
     fn get_dem_encryptor_be32<R: CryptoRngCore>(
         rng: &mut R,
-        recipient_public_key: &R25519PublicKey,
-    ) -> Result<(R25519PublicKey, aead::stream::EncryptorBE32<Aes128GcmLib>), CryptoCoreError> {
+        recipient_public_key: &NistPublicKey<C>,
+    ) -> Result<
+        (
+            NistPublicKey<C>,
+            aead::stream::EncryptorBE32<XChaCha20Poly1305Lib>,
+        ),
+        CryptoCoreError,
+    > {
         let (ephemeral_pk, key, nonce) = Self::generate_keys_and_nonce(rng, recipient_public_key)?;
-        let aes_128_gcm = Aes128Gcm::new(&key);
-        let encryptor = aes_128_gcm.into_stream_encryptor_be32(&nonce);
+        let xchacha20 = XChaCha20Poly1305::new(&key);
+        let encryptor = xchacha20.into_stream_encryptor_be32(&nonce);
         Ok((ephemeral_pk, encryptor))
     }
 
     fn get_dem_encryptor_le31<R: CryptoRngCore>(
         rng: &mut R,
-        recipient_public_key: &R25519PublicKey,
-    ) -> Result<(R25519PublicKey, aead::stream::EncryptorLE31<Aes128GcmLib>), CryptoCoreError> {
+        recipient_public_key: &NistPublicKey<C>,
+    ) -> Result<
+        (
+            NistPublicKey<C>,
+            aead::stream::EncryptorLE31<XChaCha20Poly1305Lib>,
+        ),
+        CryptoCoreError,
+    > {
         let (ephemeral_pk, key, nonce) = Self::generate_keys_and_nonce(rng, recipient_public_key)?;
-        let aes_128_gcm = Aes128Gcm::new(&key);
-        let encryptor = aes_128_gcm.into_stream_encryptor_le31(&nonce);
+        let xchacha20 = XChaCha20Poly1305::new(&key);
+        let encryptor = xchacha20.into_stream_encryptor_le31(&nonce);
         Ok((ephemeral_pk, encryptor))
     }
 
     fn get_dem_decryptor_be32(
-        recipient_private_key: &R25519PrivateKey,
-        ephemeral_public_key: &R25519PublicKey,
-    ) -> Result<aead::stream::DecryptorBE32<Aes128GcmLib>, CryptoCoreError> {
+        recipient_private_key: &NistPrivateKey<C, PRIVATE_KEY_LENGTH>,
+        ephemeral_public_key: &NistPublicKey<C>,
+    ) -> Result<aead::stream::DecryptorBE32<XChaCha20Poly1305Lib>, CryptoCoreError> {
         // recover the symmetric key and nonce
         let (key, nonce) =
             Self::recover_key_and_nonce(recipient_private_key, ephemeral_public_key)?;
         // instantiate the symmetric cipher
-        let aes_128_gcm = Aes128Gcm::new(&key);
+        let xchacha20 = XChaCha20Poly1305::new(&key);
         // turn it into a stream decryptor
-        Ok(aes_128_gcm.into_stream_decryptor_be32(&nonce))
+        Ok(xchacha20.into_stream_decryptor_be32(&nonce))
     }
 
     fn get_dem_decryptor_le31(
-        recipient_private_key: &R25519PrivateKey,
-        ephemeral_public_key: &R25519PublicKey,
-    ) -> Result<aead::stream::DecryptorLE31<Aes128GcmLib>, CryptoCoreError> {
+        recipient_private_key: &NistPrivateKey<C, PRIVATE_KEY_LENGTH>,
+        ephemeral_public_key: &NistPublicKey<C>,
+    ) -> Result<aead::stream::DecryptorLE31<XChaCha20Poly1305Lib>, CryptoCoreError> {
         // recover the symmetric key and nonce
         let (key, nonce) =
             Self::recover_key_and_nonce(recipient_private_key, ephemeral_public_key)?;
         // instantiate the symmetric cipher
-        let aes_128_gcm = Aes128Gcm::new(&key);
+        let xchacha20 = XChaCha20Poly1305::new(&key);
         // turn it into a stream decryptor
-        Ok(aes_128_gcm.into_stream_decryptor_le31(&nonce))
+        Ok(xchacha20.into_stream_decryptor_le31(&nonce))
     }
 }
 
@@ -200,30 +219,30 @@ mod tests {
 
     use super::CryptoCoreError;
     use crate::{
-        asymmetric_crypto::{R25519PrivateKey, R25519PublicKey},
-        ecies::{ecies_ristretto_aes128gcm::EciesR25519Aes128, ecies_traits::EciesStream},
+        asymmetric_crypto::{NistPrivateKey, NistPublicKey},
+        ecies::{ecies_traits::EciesStream, EciesNistChaCha20},
         reexport::rand_core::SeedableRng,
-        symmetric_crypto::Aes128Gcm,
+        symmetric_crypto::XChaCha20Poly1305,
         CsRng, Ecies, FixedSizeCBytes, RandomFixedSizeCBytes,
     };
 
     #[test]
-    fn ecies_r25519_aes128_gcm_test() {
+    fn ecies_x25519_xchacha20_poly1305_test() {
         let mut rng = CsRng::from_entropy();
         // generate a key pair
-        let private_key = R25519PrivateKey::new(&mut rng);
-        let public_key = R25519PublicKey::from(&private_key);
+        let private_key = NistPrivateKey::new(&mut rng);
+        let public_key = NistPublicKey::<C>::from(&private_key);
         // encrypt
         let plaintext = b"Hello World!";
         let ciphertext =
-            EciesR25519Aes128::encrypt(&mut rng, &public_key, plaintext, None).unwrap();
+            EciesNistChaCha20::encrypt(&mut rng, &public_key, plaintext, None).unwrap();
         // check the size is the expected size
         assert_eq!(
             ciphertext.len(),
-            plaintext.len() + EciesR25519Aes128::ENCRYPTION_OVERHEAD
+            plaintext.len() + EciesNistChaCha20::ENCRYPTION_OVERHEAD
         );
         // decrypt
-        let plaintext_ = EciesR25519Aes128::decrypt(&private_key, &ciphertext, None).unwrap();
+        let plaintext_ = EciesNistChaCha20::decrypt(&private_key, &ciphertext, None).unwrap();
         // assert
         assert_eq!(plaintext, &plaintext_[..]);
     }
@@ -231,18 +250,18 @@ mod tests {
     #[test]
     fn test_encrypt_decrypt_with_authenticated_data() -> Result<(), CryptoCoreError> {
         let mut rng = CsRng::from_entropy();
-        let private_key = R25519PrivateKey::new(&mut rng);
-        let public_key = R25519PublicKey::from(&private_key);
+        let private_key = NistPrivateKey::new(&mut rng);
+        let public_key = NistPublicKey::<C>::from(&private_key);
         let plaintext = b"Hello, World!";
         let authenticated_data = b"Optional authenticated data";
 
         // Encrypt the message
         let ciphertext =
-            EciesR25519Aes128::encrypt(&mut rng, &public_key, plaintext, Some(authenticated_data))?;
+            EciesNistChaCha20::encrypt(&mut rng, &public_key, plaintext, Some(authenticated_data))?;
 
         // Decrypt the message
         let plaintext_ =
-            EciesR25519Aes128::decrypt(&private_key, &ciphertext, Some(authenticated_data))?;
+            EciesNistChaCha20::decrypt(&private_key, &ciphertext, Some(authenticated_data))?;
 
         // Check if the decrypted message is the same as the original message
         assert_eq!(plaintext, &plaintext_[..]);
@@ -253,17 +272,17 @@ mod tests {
     #[test]
     fn test_encrypt_decrypt_with_corrupted_authentication_data() -> Result<(), CryptoCoreError> {
         let mut rng = CsRng::from_entropy();
-        let private_key = R25519PrivateKey::new(&mut rng);
-        let public_key = R25519PublicKey::from(&private_key);
+        let private_key = NistPrivateKey::new(&mut rng);
+        let public_key = NistPublicKey::<C>::from(&private_key);
         let plaintext = b"Hello, World!";
         let authenticated_data = b"Optional authenticated data";
 
         // Encrypt the message
         let ciphertext =
-            EciesR25519Aes128::encrypt(&mut rng, &public_key, plaintext, Some(authenticated_data))?;
+            EciesNistChaCha20::encrypt(&mut rng, &public_key, plaintext, Some(authenticated_data))?;
 
         // Decrypt the message
-        let not_decrypted = EciesR25519Aes128::decrypt(
+        let not_decrypted = EciesNistChaCha20::decrypt(
             &private_key,
             &ciphertext,
             Some(&b"Corrupted authenticated data"[..]),
@@ -290,11 +309,11 @@ mod tests {
         let mut rng = CsRng::from_entropy();
 
         // generate a key pair
-        let private_key = R25519PrivateKey::new(&mut rng);
-        let public_key = R25519PublicKey::from(&private_key);
+        let private_key = NistPrivateKey::new(&mut rng);
+        let public_key = NistPublicKey::<C>::from(&private_key);
 
         let (ephemeral_public_key, mut encryptor) =
-            EciesR25519Aes128::get_dem_encryptor_be32(&mut rng, &public_key)?;
+            EciesNistChaCha20::get_dem_encryptor_be32(&mut rng, &public_key)?;
 
         // prepend the ciphertext with the ephemeral public key
         let mut ciphertext = ephemeral_public_key.to_bytes().to_vec();
@@ -315,22 +334,23 @@ mod tests {
 
         //recover the ephemeral public key from the ciphertext
         let ephemeral_public_key =
-            R25519PublicKey::try_from_slice(&ciphertext[..R25519PublicKey::LENGTH])?;
+            NistPublicKey::<C>::try_from_slice(&ciphertext[..NistPublicKey::<C>::LENGTH])?;
 
         // Instantiate a decryptor
         let mut decryptor =
-            EciesR25519Aes128::get_dem_decryptor_be32(&private_key, &ephemeral_public_key)?;
+            EciesNistChaCha20::get_dem_decryptor_be32(&private_key, &ephemeral_public_key)?;
 
         // decrypt the first chunk which is BLOCK_SIZE + MAC_LENGTH bytes long
         let mut plaintext = decryptor.decrypt_next(Payload {
-            msg: &ciphertext[R25519PublicKey::LENGTH
-                ..R25519PublicKey::LENGTH + BLOCK_SIZE + Aes128Gcm::MAC_LENGTH],
+            msg: &ciphertext[NistPublicKey::<C>::LENGTH
+                ..NistPublicKey::<C>::LENGTH + BLOCK_SIZE + XChaCha20Poly1305::MAC_LENGTH],
             aad,
         })?;
 
         // decrypt the second and last chunk
         plaintext.extend_from_slice(&decryptor.decrypt_last(Payload {
-            msg: &ciphertext[R25519PublicKey::LENGTH + BLOCK_SIZE + Aes128Gcm::MAC_LENGTH..],
+            msg: &ciphertext
+                [NistPublicKey::<C>::LENGTH + BLOCK_SIZE + XChaCha20Poly1305::MAC_LENGTH..],
             aad,
         })?);
 
@@ -353,11 +373,11 @@ mod tests {
         let mut rng = CsRng::from_entropy();
 
         // generate a key pair
-        let private_key = R25519PrivateKey::new(&mut rng);
-        let public_key = R25519PublicKey::from(&private_key);
+        let private_key = NistPrivateKey::new(&mut rng);
+        let public_key = NistPublicKey::<C>::from(&private_key);
 
         let (ephemeral_public_key, mut encryptor) =
-            EciesR25519Aes128::get_dem_encryptor_le31(&mut rng, &public_key)?;
+            EciesNistChaCha20::get_dem_encryptor_le31(&mut rng, &public_key)?;
 
         // prepend the ciphertext with the ephemeral public key
         let mut ciphertext = ephemeral_public_key.to_bytes().to_vec();
@@ -378,22 +398,23 @@ mod tests {
 
         //recover the ephemeral public key from the ciphertext
         let ephemeral_public_key =
-            R25519PublicKey::try_from_slice(&ciphertext[..R25519PublicKey::LENGTH])?;
+            NistPublicKey::<C>::try_from_slice(&ciphertext[..NistPublicKey::<C>::LENGTH])?;
 
         // Instantiate a decryptor
         let mut decryptor =
-            EciesR25519Aes128::get_dem_decryptor_le31(&private_key, &ephemeral_public_key)?;
+            EciesNistChaCha20::get_dem_decryptor_le31(&private_key, &ephemeral_public_key)?;
 
         // decrypt the first chunk which is BLOCK_SIZE + MAC_LENGTH bytes long
         let mut plaintext = decryptor.decrypt_next(Payload {
-            msg: &ciphertext[R25519PublicKey::LENGTH
-                ..R25519PublicKey::LENGTH + BLOCK_SIZE + Aes128Gcm::MAC_LENGTH],
+            msg: &ciphertext[NistPublicKey::<C>::LENGTH
+                ..NistPublicKey::<C>::LENGTH + BLOCK_SIZE + XChaCha20Poly1305::MAC_LENGTH],
             aad,
         })?;
 
         // decrypt the second and last chunk
         plaintext.extend_from_slice(&decryptor.decrypt_last(Payload {
-            msg: &ciphertext[R25519PublicKey::LENGTH + BLOCK_SIZE + Aes128Gcm::MAC_LENGTH..],
+            msg: &ciphertext
+                [NistPublicKey::<C>::LENGTH + BLOCK_SIZE + XChaCha20Poly1305::MAC_LENGTH..],
             aad,
         })?);