+++ src/ecies/mod.rs
@@ -1,45 +1,37 @@
-#[cfg(all(
-    feature = "ecies",
-    feature = "aes",
-    feature = "sha3",
-    feature = "curve25519"
-))]
-mod ecies_ristretto_aes128gcm;
-#[cfg(all(
-    feature = "ecies",
-    feature = "chacha",
-    feature = "blake",
-    feature = "curve25519"
-))]
-mod ecies_salsa_sealed_box;
-mod ecies_traits;
-#[cfg(all(
-    feature = "ecies",
-    feature = "chacha",
-    feature = "blake",
-    feature = "curve25519"
-))]
-mod ecies_x25519_xchacha20;
+mod traits;
+pub use traits::*;
+
+#[cfg(all(feature = "ecies", feature = "aes", feature = "sha3"))]
+mod generic_ecies_aes128gcm;
+#[cfg(all(feature = "ecies", feature = "aes", feature = "sha3"))]
+pub use generic_ecies_aes128gcm::EciesAes128;
 
 #[cfg(all(
     feature = "ecies",
-    feature = "aes",
+    feature = "curve25519",
+    feature = "chacha",
+    feature = "blake"
+))]
+mod curve_25519;
+#[cfg(all(
+    feature = "ecies",
+    feature = "curve25519",
+    feature = "chacha",
+    feature = "blake"
+))]
+pub use curve_25519::*;
+
+#[cfg(all(
+    feature = "ecies",
+    feature = "nist_curves",
     feature = "sha3",
-    feature = "curve25519"
+    feature = "aes"
 ))]
-pub use ecies_ristretto_aes128gcm::EciesR25519Aes128;
+mod nist_curves;
 #[cfg(all(
     feature = "ecies",
-    feature = "chacha",
-    feature = "blake",
-    feature = "curve25519"
+    feature = "nist_curves",
+    feature = "sha3",
+    feature = "aes"
 ))]
-pub use ecies_salsa_sealed_box::EciesSalsaSealBox;
-pub use ecies_traits::{Ecies, EciesStream};
-#[cfg(all(
-    feature = "ecies",
-    feature = "chacha",
-    feature = "blake",
-    feature = "curve25519"
-))]
-pub use ecies_x25519_xchacha20::EciesX25519XChaCha20;
+pub use nist_curves::*;